requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
numpy==2.2.0
pillow==11.1.0
psycopg2-binary==2.9.10
pyarrow==18.1.0
PyJWT==2.10.1
pylance==0.20.0
python-dotenv==1.0.1
pytz==2024.2
sqlparse==0.5.2
tzdata==2024.2
